feat(gfx): 添加洋葱皮渲染支持
- 新增 onion_note.wgsl 着色器，实现半透明（alpha=0.3）无边框音符渲染
- 在 NoteRenderer 中添加洋葱皮专用初始化方法和着色器支持
- 将 NoteInstance 的 key 参数类型从 f32 改为 u8，统一编码方式
- 实现洋葱皮状态管理和全量实例构建逻辑，支持音轨开关过滤
- 集成洋葱皮渲染到主渲染管线，置于网格和主音轨之间
- 添加相关单元测试验证洋葱皮状态机行为

+++ crates/ui-editor/src/note.rs
@@ -35,7 +35,7 @@ impl NoteExt for Note {
     fn to_instance(&self, color: Color, border_width: u32) -> NoteInstance {
         NoteInstance::new(
             self.tick,
-            self.key as f32,
+            self.key as u8,
             self.length,
             color_to_array(color),
             border_width,
